extract k8cluster receiver logic from kubeletstats_receiver.go
